pyproject.toml
[project]
name = "code-reviewer"
version = "0.1.0"
description = "AI-powered code review tool"
authors = [{ name = "Gerrit Vermeulen" }]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "click>=8.1",
    "google-adk>=1.18.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.4",
]

[project.scripts]
code-reviewer = "code_reviewer.cli:main"

[tool.uv]
package = true
